accident_report/requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
sqlparse==0.5.3
tzdata==2025.1
psycopg2-binary>=2.9.10
python-dotenv
django-widget-tweaks
